+++ Changes.md
@@ -1,7 +1,8 @@
 ## v3.4 (unreleased)
 
-* Remote: Select new sprite by right-clicking sprite button
-* Display current sprite preview image on button (configurable)
+* Remote: Display current sprite preview image on button
+* Select new sprite by right-clicking sprite button
+* Image or message tooltip information depending on set mode
 * Fix: Audio capture not working on some systems (#16)
 * Fix: Inconsistent display of switched audio device